feat(arc-201): allow bold text in notification description
